# web/src/lib/generated/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "SkillPass — skills-based hiring platform API",
        "title": "SkillPass API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/hris/documents": {
            "get": {
                "tags": [
                    "documents"
                ],
                "summary": "List documents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DocumentResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "documents"
                ],
                "summary": "Upload a document",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "description": "File (max 25MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "category": {
                                        "description": "identity|contract|certificate|payslip|tax|other",
                                        "type": "string"
                                    },
                                    "employeeId": {
                                        "description": "Employee to attach the document to",
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/hris/documents/audit-log": {
            "get": {
                "tags": [
                    "documents"
                ],
                "summary": "Document access audit log",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DocumentAccessLog"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/hris/documents/{id}": {
            "delete": {
                "tags": [
                    "documents"
                ],
                "summary": "Delete a document",
                "parameters": [
                    {
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {},
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/hris/documents/{id}/download": {
            "get": {
                "tags": [
                    "documents"
                ],
                "summary": "Download a document",
                "parameters": [
                    {
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {},
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "//localhost:1234/api/v1"
        }
    ],
    "components": {
...
        "securitySchemes": {
            "BearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "DocumentAccessLog": {
                "type": "object",
                "properties": {
                    "accessedByName": {
                        "type": "string"
                    },
                    "action": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "documentId": {
                        "type": "string"
                    },
                    "filename": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "ipAddress": {
                        "type": "string"
                    }
                }
            },
            "DocumentResponse": {
                "type": "object",
                "properties": {
                    "category": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "employeeId": {
                        "type": "string"
                    },
                    "employeeName": {
                        "type": "string"
                    },
                    "fileSize": {
                        "type": "integer"
                    },
                    "filename": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "mimeType": {
                        "type": "string"
                    },
                    "scanStatus": {
                        "type": "string"
                    },
                    "uploadedByName": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}